+++ include/placeholders_interface/Placeholders.h
@@ -48,4 +48,5 @@ public:
     static std::string GetFileIdentity(const std::wstring &filePath);
     static void UpdateFileIdentity(const std::wstring &filePath, const std::wstring &fileIdentity, bool isDirectory);
     static PlaceholderAttribute GetAttribute(const std::wstring &filePath);
+    static FileState GetPlaceholderInfo(const std::wstring &directoryPath);
 };
