+++ include/sync_root_interface/Utilities.h
@@ -10,6 +10,7 @@ public:
     static void ApplyCustomOverwriteStateToPlaceholderFile(_In_ LPCWSTR path, _In_ LPCWSTR filename, _In_ winrt::StorageProviderItemProperty &prop);
     static std::wstring ProcessErrorNameToWString(_In_ ProcessErrorName error);
     static std::wstring FileOperationErrorToWString(_In_ FileOperationError error);
+    static bool IsTemporaryFile(const std::wstring &fullPath);
 
     static winrt::com_array<wchar_t>
     ConvertSidToStringSid(_In_ PSID sid)
