+++ include/sync_root_interface/callbacks/Callbacks.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <CallbacksContext.h>
-#include "DirectoryWatcher.h"
 
 void register_threadsafe_callbacks(napi_env env, InputSyncCallbacks input);
 
@@ -24,11 +23,3 @@ void CALLBACK fetch_data_callback_wrapper(_In_ CONST CF_CALLBACK_INFO *callbackI
 // Fetch Data Cancel Callback
 void register_threadsafe_cancel_fetch_data_callback(const std::string &resource_name, napi_env env, InputSyncCallbacks input);
 void CALLBACK cancel_fetch_data_callback_wrapper(_In_ CONST CF_CALLBACK_INFO *callbackInfo, _In_ CONST CF_CALLBACK_PARAMETERS *callbackParameters);
-
-// Notify File Added Callback. This is a fake callback
-void notify_file_added_call(napi_env env, napi_value js_callback, void *context, void *data);
-void register_threadsafe_notify_file_added_callback(FileChange &change, const std::string &resource_name, napi_env env, InputSyncCallbacksThreadsafe input);
-
-// Notify message from node-win
-void notify_message_call(napi_env env, napi_value js_callback, void *context, void *data);
-void register_threadsafe_message_callback(FileChange &change, const std::string &resource_name, napi_env env, InputSyncCallbacksThreadsafe input);
